refactor JNA package
