+++ src/main/java/bh/bot/app/AbstractApplication.java
@@ -42,6 +42,8 @@ import bh.bot.common.Configuration;
 import bh.bot.common.Configuration.Offset;
 import bh.bot.common.Telegram;
 import bh.bot.common.exceptions.InvalidDataException;
+import bh.bot.common.jna.IJna;
+import bh.bot.common.jna.SteamWindowsJna;
 import bh.bot.common.types.ParseArgumentsResult;
 import bh.bot.common.types.ScreenResolutionProfile;
 import bh.bot.common.types.annotations.AppCode;
@@ -49,8 +51,6 @@ import bh.bot.common.types.flags.FlagPattern;
 import bh.bot.common.types.flags.FlagResolution;
 import bh.bot.common.types.flags.Flags;
 import bh.bot.common.types.images.BwMatrixMeta;
-import bh.bot.common.types.jna.IJna;
-import bh.bot.common.types.jna.SteamWindowsJna;
 import bh.bot.common.types.tuples.Tuple2;
 import bh.bot.common.types.tuples.Tuple3;
 import bh.bot.common.types.tuples.Tuple4;
