+++ src/main/java/bh/bot/app/dev/TestApp.java
@@ -3,11 +3,11 @@ package bh.bot.app.dev;
 import bh.bot.app.AbstractApplication;
 import bh.bot.common.Configuration;
 import bh.bot.common.Configuration.Offset;
+import bh.bot.common.jna.IJna;
+import bh.bot.common.jna.SteamWindowsJna;
 import bh.bot.common.types.AttendablePlace;
 import bh.bot.common.types.AttendablePlaces;
 import bh.bot.common.types.annotations.AppCode;
-import bh.bot.common.types.jna.IJna;
-import bh.bot.common.types.jna.SteamWindowsJna;
 import bh.bot.common.types.tuples.Tuple2;
 import bh.bot.common.types.tuples.Tuple4;
 import bh.bot.common.utils.InteractionUtil;
