+++ src/main/java/bh/bot/common/jna/AbstractWindowsJna.java
@@ -1,4 +1,4 @@
-package bh.bot.common.types.jna;
+package bh.bot.common.jna;
 
 import com.sun.jna.platform.win32.User32;
